feat(svg): use IBM Plex Mono for background text
Summary:

Test Plan:
